Fix some warnings in pco.cpp

+++ pcowinApp/src/Pco.cpp
@@ -384,7 +384,7 @@ void Pco::initialiseOnceRunning()
 
 /**
  * Return the pco corresponding to the port name
- * \param[in] p The port name
+ * \param[in] portName The port name
  * \return The pco object, NULL if not found
  */
 Pco* Pco::getPco(const char* portName)
@@ -1709,7 +1709,7 @@ void Pco::post(const StateMachine::Event* req)
  */
 NDArray* Pco::allocArray(int sizeX, int sizeY, NDDataType_t dataType)
 {
-    size_t maxDims[] = {sizeX, sizeY};
+    size_t maxDims[] = {(size_t) sizeX, (size_t) sizeY};
     NDArray* image = this->pNDArrayPool->alloc(sizeof(maxDims)/sizeof(size_t),
             maxDims, dataType, 0, NULL);
     if(image == NULL)
